Motivation:

Existing Recycler has bad performance in the thread-local case

Modification:

Allow Recycler to drop fast-thread-local usage and enable fully unguarded recycles i.e. no wrapper handling

Result:

Fixes #15604

- [x] Implement correctly https://github.com/netty/netty/pull/15610#issuecomment-3258147192
- [x] Add tests for missing recycler types
- [x] Implementing a mechanism to allow pinned recyclers to gc their owner + tests
- [ ] Improve performance for the #15419