@franz1981 2025-09-03
Some benchmark numbers:
```
Benchmark                                                                 (fastThreadLocal)  (unguarded)  Mode  Cnt    Score   Error      Units
RecyclerBenchmark.recyclerGetAndUnguardedRecycle                                      false         true  avgt   20    3.697 ± 0.031      ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:CPI                                  false         true  avgt    2    0.210          clks/insn
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:IPC                                  false         true  avgt    2    4.755          insns/clk
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-dcache-load-misses                false         true  avgt    2    0.003               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-dcache-loads                      false         true  avgt    2   32.647               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-icache-load-misses                false         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-icache-loads                      false         true  avgt    2    0.043               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:branch-misses                        false         true  avgt    2    0.004               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:branches                             false         true  avgt    2   19.577               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:cycles                               false         true  avgt    2   19.990               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:dTLB-load-misses                     false         true  avgt    2   ≈ 10⁻⁵               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:dTLB-loads                           false         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:iTLB-load-misses                     false         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:iTLB-loads                           false         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:instructions                         false         true  avgt    2   95.068               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:stalled-cycles-frontend              false         true  avgt    2    0.104               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle                                       true         true  avgt   20    4.430 ± 0.072      ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:CPI                                   true         true  avgt    2    0.218          clks/insn
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:IPC                                   true         true  avgt    2    4.587          insns/clk
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-dcache-load-misses                 true         true  avgt    2    0.004               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-dcache-loads                       true         true  avgt    2   38.730               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-icache-load-misses                 true         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:L1-icache-loads                       true         true  avgt    2    0.051               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:branch-misses                         true         true  avgt    2    0.005               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:branches                              true         true  avgt    2   23.635               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:cycles                                true         true  avgt    2   24.029               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:dTLB-load-misses                      true         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:dTLB-loads                            true         true  avgt    2   ≈ 10⁻³               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:iTLB-load-misses                      true         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:iTLB-loads                            true         true  avgt    2   ≈ 10⁻⁴               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:instructions                          true         true  avgt    2  110.240               #/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle:stalled-cycles-frontend               true         true  avgt    2    0.131               #/op
```
vs a fully local one:
```
Benchmark                                                             (unguarded)  Mode  Cnt   Score   Error      Units
RecyclerBenchmark.localRecyclerGetAndRecycle                                false  avgt   20   2.992 ± 0.011      ns/op
RecyclerBenchmark.localRecyclerGetAndRecycle:CPI                            false  avgt    2   0.221          clks/insn
RecyclerBenchmark.localRecyclerGetAndRecycle:IPC                            false  avgt    2   4.521          insns/clk
RecyclerBenchmark.localRecyclerGetAndRecycle:L1-dcache-load-misses          false  avgt    2   0.001               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:L1-dcache-loads                false  avgt    2  26.143               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:L1-icache-load-misses          false  avgt    2  ≈ 10⁻⁴               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:L1-icache-loads                false  avgt    2   0.021               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:branch-misses                  false  avgt    2   0.001               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:branches                       false  avgt    2  13.074               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:cycles                         false  avgt    2  16.008               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:dTLB-load-misses               false  avgt    2  ≈ 10⁻⁵               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:dTLB-loads                     false  avgt    2  ≈ 10⁻⁴               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:iTLB-load-misses               false  avgt    2  ≈ 10⁻⁵               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:iTLB-loads                     false  avgt    2  ≈ 10⁻⁴               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:instructions                   false  avgt    2  72.378               #/op
RecyclerBenchmark.localRecyclerGetAndRecycle:stalled-cycles-frontend        false  avgt    2   0.050               #/op
```
The number of instruction for the `fastThreadLocal = false unguarded = true` case is still higher than what we have in the local case, but I think that it depends by the additional checks and re-read of fields caused by the volatile fields in the LocalPool. I have to check if there's something we can do about it.

Let me know @georgebanasios if you can use the non fast thread local and unguarded version of `Recycler` for the Adaptive PR.
Using it it should be as simple as `new Recycler(Thread ownerThread, boolean unguarded)` stored in the magazine's field (or magazine group if we know is both pinned and uncontended): we're now ~25% slower than it.