@franz1981 2025-09-03
I have to re-run the numbers of the `fastThreadLocal = true` case since we now add few more checks (I could simplify one of them).