@franz1981 2025-09-03
So, to summarize

before these changes, JDK 21 (and **JDK 24 too**):
```
Benchmark                                         Mode  Cnt  Score   Error  Units
RecyclerBenchmark.recyclerGetAndUnguardedRecycle  avgt   20  5.510 ± 0.021  ns/op
```
at https://github.com/netty/netty/commit/f2726c870a8e33ce460fee43af0dc94622861ab2 - with JDK 21:
```
Benchmark                                         (fastThreadLocal)  (unguarded)  Mode  Cnt  Score   Error  Units
# this can be compared to the original one
RecyclerBenchmark.recyclerGetAndUnguardedRecycle               true        false  avgt   20  5.189 ± 0.025  ns/op
# these are all additional ways
RecyclerBenchmark.recyclerGetAndUnguardedRecycle              false         true  avgt   20  3.622 ± 0.013  ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle              false        false  avgt   20  4.390 ± 0.124  ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle               true         true  avgt   20  4.458 ± 0.008  ns/op
```
Whilst with JDK 24:
```
Benchmark                                         (fastThreadLocal)  (unguarded)  Mode  Cnt  Score   Error  Units
RecyclerBenchmark.recyclerGetAndUnguardedRecycle               true        false  avgt   20  5.634 ± 0.206  ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle              false         true  avgt   20  3.317 ± 0.013  ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle              false        false  avgt   20  3.968 ± 0.007  ns/op
RecyclerBenchmark.recyclerGetAndUnguardedRecycle               true         true  avgt   20  5.622 ± 0.011  ns/op
```
The new version has a severe performance hit (still within original performance) regardless been `unguarded` - which is surprising - and which seems dependent by `fastThreadLocal = true` only.